Move MatchPredictionListAdapter to :prediction
