+++ prediction/src/main/res/drawable/ic_edit_white_24dp.xml
