+++ prediction/src/main/res/drawable/ic_nha_white_24dp.xml
