+++ prediction/src/main/res/drawable/ic_rwc_white_24dp.xml
