+++ prediction/src/main/res/layout/list_item_match_prediction.xml
