+++ prediction/src/main/res/values/strings.xml
@@ -15,6 +15,8 @@
     <string name="check_box_rugby_world_cup">Rugby World Cup</string>
     <string name="check_box_no_home_advantage">No home advantage (neutral venue)</string>
 
+    <string name="chip_match_prediction" translatable="false">%1$s %2$d - %3$d %4$s</string>
+
     <string name="content_description_close">Close</string>
 
 </resources>
